Fix warnings from code check tools
* Fixed shellcheck warnings

+++ remove_modules.sh
@@ -1,15 +1,15 @@
-#!/bin/sh
+#!/bin/sh -e
 # remove all external puppet modules
 
 PLUGIN_MOD_DIR="deployment_scripts/puppet/modules"
 
-dir=`dirname $0`
+dir=$(dirname $0)
 cd "${dir}" || exit 1
 
-cat Puppetfile | grep "^mod '" | awk -F "'" '{ print $2 }' | while read module; do
+grep "^mod '" Puppetfile| awk -F "'" '{ print $2 }' | while read module; do
   if [ -d "${PLUGIN_MOD_DIR}/${module}" ]; then
     echo "Remove: ${PLUGIN_MOD_DIR}/${module}"
-    rm -rf "${PLUGIN_MOD_DIR}/${module}"
+    rm -rf "${PLUGIN_MOD_DIR:?}/${module}"
   fi
 done
 rm -f 'Puppetfile.lock'
